Buat direktori baru untuk fixture tes
