+++ tests/test_str.py
@@ -95,7 +95,7 @@ def test_menjadikan(menjadikan):
     assert str(menjadikan) == result
 
 
-def test_lampir(lampir):
+def test_prakategorial_lampir(lampir):
     result = (
         "lam.pir\n"
         "(prakategorial)  cari: lampiran, melampiri, melampirkan, terlampir"
@@ -125,7 +125,7 @@ def test_kan(kan):
     assert str(kan) == result
 
 
-def test_me_(me_):
+def test_awalan_me(awalan_me):
     result = (
         "me-  /mê-/\n"
         "→ meng-\n"
@@ -143,22 +143,22 @@ def test_me_(me_):
         "9. (prefiks pembentuk verba)  membuat; menghasilkan: menyambal; menggulai; membatik\n"
         "10. (prefiks pembentuk verba)  menyatakan: mengaku"
     )
-    assert str(me_) == result
+    assert str(awalan_me) == result
 
 
-def test__kan(_kan):
+def test_akhiran_kan(akhiran_kan):
     result = (
         "-kan [5]\n"
         "1. (sufiks pembentuk verba)  menjadikan: jalankan; datangkan; hitamkan\n"
         "2. (sufiks pembentuk verba)  sungguh-sungguh: dengarkan; camkan\n"
         "3. (sufiks pembentuk verba)  untuk; kepada orang lain: sewakan; bacakan"
     )
-    assert str(_kan) == result
+    assert str(akhiran_kan) == result
 
 
-def test__lah(_lah):
+def test_terikat_lah(terikat_lah):
     result = (
         "-lah [1]\n"
         "(bentuk terikat)  yang digunakan untuk menekankan makna kata yang di depannya"
     )
-    assert str(_lah) == result
+    assert str(terikat_lah) == result
